cr_tests/libcrtest: make u64 unsigned long long
...which is 64 bits on all platforms and helps us avoid warnings like

libcrtest/common.c:374: warning: format '%llu' expects type 'long long unsigned int', but argument 5 has type 'u64'

which would otherwise need tiresome (unsigned long long) casts.

+++ libcrtest/libcrtest.h
@@ -12,8 +12,7 @@ struct record {
 	char data[256];
 };
 
-#include <stdint.h>
-typedef uint64_t u64;
+typedef unsigned long long u64;
 
 extern void do_exit(int status);
 extern int test_done(void);
